Worland transform tests with triangular truncation (#587)
- Change splitting metadata to include 1D truncation for each mode (needed for triangular truncation)
- Drop short vs full metadata setup and only store full data
- Update all LoadSplitter tests with new metadata format
- Add triangular truncation LoadSplitter tests
- Update metadata in Worland tests
- Provide metadata for triangular Worland tests

+++ Components/Transform/TestSuite/Scripts/generate_meta_split_worland.sh
@@ -1,52 +1,70 @@
 #!/bin/bash
+#
+# RUN FROM build/Components/Transform/TestSuite
 
 scheme="WLFl"
 stage=0
+
+cleardata=true
+copy2Ref=true
+
+smallDbs=( 104 )
+bigDbs=( 108 109 110 111 )
 
 data_dir="_data/Framework/LoadSplitter/${scheme}/Tubular"
 ref_dir="_refdata/Transform/Worland/Integrator/"
 exe="../../Framework/TestSuite/FrameworkLoadSplitterTests"
 
+logfile="generate_meta_split_worland.log"
+
 # Create path if it doesn't exist
 mkdir -p "${data_dir}"
 mkdir -p "${ref_dir}"
 
-# Clean data
-rm "./${data_dir}"/*
+# Clear data
+if [ "$clearData" = true ];
+then
+  echo "Clearing old data"
+  rm "./${data_dir}"/*
+fi
 
 # DB 104
-db=104
-for np in 4;
+for db in ${smallDbs[@]};
 do
-  for r in 0 1 $((${np} -1));
+  for np in 4 8;
   do
-    ${exe} \[${scheme}\] --algorithm tubular --np ${np} --db ${db} --stage ${stage} --id ${r} --dumpData
-  done
-done
-for np in 8;
-do
-  for r in 0 3 $((${np} -1));
-  do
-    ${exe} \[${scheme}\] --algorithm tubular --np ${np} --db ${db} --stage ${stage} --id ${r} --dumpData
+    for r in 0 $(( (${np} - 1)/2 )) $(( ${np} - 1 ));
+    do
+      echo "Generating metadata for ${db}: np = ${db}, r = ${r}"
+      ${exe} \[${scheme}\] --algorithm tubular --np ${np} --db ${db} --stage ${stage} --id ${r} --dumpData > ${logfile}
+    done
   done
 done
 
 # DB 108, 109, 110, 111
-for db in 108 109 110 111;
+for db in ${bigDbs[@]};
 do
   for n in {0..10};
   do
     for np in $(( 2**n )) $(( 2**n*12 )) $(( 2**n*18 ));
     do
-      if [ "$np" -gt "7" ];
+      if [ "$np" -gt "3" ];
       then
-        for r in 0 3 7 $((${np}-1));
+        for r in 0 3 7 $(( ${np} - 1 ));
         do
-          ${exe} \[${scheme}\] --algorithm tubular --np ${np} --db ${db} --stage ${stage} --id ${r} --dumpData
+          if [ "$np" -gt "$r" ];
+          then
+            echo "Generating metadata for ${db}: np = ${db}, r = ${r}"
+            ${exe} \[${scheme}\] --algorithm tubular --np ${np} --db ${db} --stage ${stage} --id ${r} --dumpData > ${logfile}
+          fi
         done
       fi
     done
   done
 done
 
-#cp -r "./${data_dir}"/* "./${ref_dir}"/
+if [ "$copy2Ref" = true ];
+then
+  echo "Copying new data to ref"
+  cp -r "./${data_dir}"/* "./${ref_dir}"/
+fi
